Add files via upload
This was my first project as a web developer.
I am not going to lie here and pretend that I built everything from scratch, no!
I customized another project that I saw on the internet. I found the colors and the design too simplistic, so I decided that I wanted to see the 
source code and try my htlm, css and javascript skills.

I am open to suggestions to improve this calculator and make it more modern.